Move ingestAcceptedStatus to IngestJobStatusFromJobTestData
