+++ java/ingest/ingest-core/src/test/java/sleeper/ingest/core/job/status/IngestJobStatusInPeriodTest.java
@@ -226,7 +226,7 @@ public class IngestJobStatusInPeriodTest {
     }
 
     private ProcessStatusUpdate startedRun(Instant startedTime) {
-        return IngestJobStatusTestHelper.ingestStartedStatus(job, startedTime, defaultUpdateTime(startedTime));
+        return IngestJobStatusTestHelper.ingestStartedStatus(startedTime, defaultUpdateTime(startedTime));
     }
 
     private ProcessStatusUpdate finishedRun(Instant startedTime, Instant finishedTime) {
